회원가입, 로그인, 삭제 API Response의 DTO를 ApiResponseDto로 통합 JwtUtil에  게시글 작성, 수정, 삭제 시 토큰을 검사하고, 토큰을 복호화하여 username을 가져오는 메서드 추가 (-> 위험한 방식인지 모르겠음)
게시글 관련 API 수정:
인증 방식을 JWT로 변경, 기존 Post Entity의 userName, password 필드 제거, 게시글 작성, 조회, 수정, 삭제 메서드 수정

Database post -> blog 변경

Post:User = N:1, Post @ManyToOne 단방향 관계 설정

+++ src/main/java/com/sparta/springBlog/service/PostService.java
@@ -2,23 +2,30 @@ package com.sparta.springBlog.service;
 
 import com.sparta.springBlog.dto.PostRequestDto;
 import com.sparta.springBlog.dto.PostResponseDto;
+import com.sparta.springBlog.dto.ApiResponseDto;
 import com.sparta.springBlog.entity.Post;
+import com.sparta.springBlog.entity.User;
 import com.sparta.springBlog.repository.PostRepository;
+import com.sparta.springBlog.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
 @Service
+@RequiredArgsConstructor
 public class PostService {
     private final PostRepository postRepository;
+    private final UserRepository userRepository;
 
-    public PostService(PostRepository postRepository) {
-        this.postRepository = postRepository;
-    }
-
-    public PostResponseDto createPost(PostRequestDto postRequestDto) {
-        Post post = new Post(postRequestDto);
+    public PostResponseDto createPost(PostRequestDto postRequestDto, String userName) {
+        Post post = new Post(postRequestDto, userName);
+        User user = userRepository.findByUsername(userName).orElseThrow(() ->
+                new IllegalArgumentException("userName이 잘못되었습니다.") // 로직 상 userName 은 이미 검증이 된 상태이기에 예외가 발생하는 경우는 없다
+        );
+        post.setUser(user);
 
         Post savePost = postRepository.save(post);
 
@@ -36,31 +43,28 @@ public class PostService {
     }
 
     @Transactional
-    public PostResponseDto updatePost(Long id, PostRequestDto postRequestDto) {
-        Post post = findPost(id, postRequestDto.getPassword());
-
-        post.update(postRequestDto);
-
+    public PostResponseDto updatePost(Long id, PostRequestDto postRequestDto, String userName) {
+        Post post = findPost(id);
+        if (!post.getUserName().equals(userName)) {
+            throw new IllegalArgumentException("userName이 일치하지 않습니다.");
+        }
+        post.setPostName(postRequestDto.getPostName());
+        post.setPostContent(postRequestDto.getPostContent());
         return new PostResponseDto(post);
     }
 
-    public PostResponseDto deletePost(Long id, PostRequestDto postRequestDto) {
-        Post post = findPost(id, postRequestDto.getPassword());
-
+    public ApiResponseDto deletePost(Long id, String userName) {
+        Post post = findPost(id);
+        if (!post.getUserName().equals(userName)) {
+            throw new IllegalArgumentException("userName이 일치하지 않습니다.");
+        }
         postRepository.delete(post);
-
-        return new PostResponseDto(true);
+        return new ApiResponseDto("게시글 삭제 성공", HttpStatus.OK.value());
     }
 
     private Post findPost(Long id) {
         return postRepository.findById(id).orElseThrow(() ->
                 new IllegalArgumentException("해당 ID의 게시글이 존재하지 않습니다.")
-        );
-    }
-
-    private Post findPost(Long id, String password) {
-        return postRepository.findByIdAndPassword(id, password).orElseThrow(() ->
-                new IllegalArgumentException("ID 혹은 비밀번호 오류입니다.")
         );
     }
 }
